+++ rbxmk/testdata/rbxmk/types/MemberDesc_members.lua
@@ -1,8 +1,8 @@
 -- Test members common to member descriptors.
-local prop = rbxmk.newDesc("PropertyDesc")
-local func = rbxmk.newDesc("FunctionDesc")
-local event = rbxmk.newDesc("EventDesc")
-local callback = rbxmk.newDesc("CallbackDesc")
+local prop = PropertyDesc.new()
+local func = FunctionDesc.new()
+local event = EventDesc.new()
+local callback = CallbackDesc.new()
 
 -- Parameters
 for _, desc in ipairs({func,event,callback}) do
@@ -15,41 +15,38 @@ for _, desc in ipairs({func,event,callback}) do
 	T.Fail(function() desc:SetParameters(42) end    , t..": cannot call SetParameters method with non-table")
 	T.Fail(function() desc:SetParameters() end      , t..": cannot call SetParameters method with nil")
 	T.Fail(function()
-		desc:SetParameters(rbxmk.newDesc("ParameterDesc"))
+		desc:SetParameters(ParameterDesc.new())
 	end, t..": cannot call SetParameters method with ParameterDesc")
 	T.Pass(function()
 		desc:SetParameters({
-			rbxmk.newDesc("ParameterDesc", rbxmk.newDesc("TypeDesc", "FooCatA", "FooTypeA"), "fooNameA"),
-			rbxmk.newDesc("ParameterDesc", rbxmk.newDesc("TypeDesc", "FooCatB", "FooTypeB"), "fooNameB", "FooDefault"),
-			rbxmk.newDesc("ParameterDesc", rbxmk.newDesc("TypeDesc", "FooCatC", "FooTypeC"), "fooNameC"),
+			ParameterDesc.new(TypeDesc.new("FooCatA", "FooTypeA"), "fooNameA"),
+			ParameterDesc.new(TypeDesc.new("FooCatB", "FooTypeB"), "fooNameB", "FooDefault"),
+			ParameterDesc.new(TypeDesc.new("FooCatC", "FooTypeC"), "fooNameC"),
 		})
 	end, t..": can call SetParameters method with table of ParameterDescs")
 	T.Fail(function()
 		desc:SetParameters({
-			rbxmk.newDesc("ParameterDesc", rbxmk.newDesc("TypeDesc", "FooCatA", "FooTypeA"), "fooNameA"),
+			ParameterDesc.new(TypeDesc.new("FooCatA", "FooTypeA"), "fooNameA"),
 			"Foobar",
-			rbxmk.newDesc("ParameterDesc", rbxmk.newDesc("TypeDesc", "FooCatC", "FooTypeC"), "fooNameC"),
+			ParameterDesc.new(TypeDesc.new("FooCatC", "FooTypeC"), "fooNameC"),
 		})
 	end, t..": cannot call SetParameters method with table of non-ParameterDescs")
 	T.Pass(function()
-		return desc:Parameters()[1] == rbxmk.newDesc(
-			"ParameterDesc",
-			rbxmk.newDesc("TypeDesc", "FooCatA", "FooTypeA"),
+		return desc:Parameters()[1] == ParameterDesc.new(
+			TypeDesc.new("FooCatA", "FooTypeA"),
 			"fooNameA"
 		)
 	end, t..": first set parameter persists")
 	T.Pass(function()
-		return desc:Parameters()[2] == rbxmk.newDesc(
-			"ParameterDesc",
-			rbxmk.newDesc("TypeDesc", "FooCatB", "FooTypeB"),
+		return desc:Parameters()[2] == ParameterDesc.new(
+			TypeDesc.new("FooCatB", "FooTypeB"),
 			"fooNameB",
 			"FooDefault"
 		)
 	end, t..": second set parameter persists")
 	T.Pass(function()
-		return desc:Parameters()[3] == rbxmk.newDesc(
-			"ParameterDesc",
-			rbxmk.newDesc("TypeDesc", "FooCatC", "FooTypeC"),
+		return desc:Parameters()[3] == ParameterDesc.new(
+			TypeDesc.new("FooCatC", "FooTypeC"),
 			"fooNameC"
 		)
 	end, t..": third set parameter persists")
@@ -62,9 +59,9 @@ for _, desc in ipairs({func,callback}) do
 		t..": can get ReturnType field")
 	T.Pass(function() return typeof(desc.ReturnType) == "TypeDesc" end,
 		t..": ReturnType field is a TypeDesc")
-	T.Pass(function() return desc.ReturnType == rbxmk.newDesc("TypeDesc") end,
+	T.Pass(function() return desc.ReturnType == TypeDesc.new() end,
 		t..": ReturnType field initializes to empty TypeDesc")
-	T.Pass(function() desc.ReturnType = rbxmk.newDesc("TypeDesc", "FooCategory", "FooName") end,
+	T.Pass(function() desc.ReturnType = TypeDesc.new("FooCategory", "FooName") end,
 		t..": can set ReturnType field to TypeDesc")
 	T.Fail(function() desc.ReturnType = 42 end,
 		t..": cannot set ReturnType field to non-string")